requirements.txt
# Core libraries
requests>=2.28.0
pandas>=1.5.0
numpy>=1.23.0

# Database
psycopg2-binary>=2.9.0

# Dashboard & Visualization
streamlit>=1.20.0
plotly>=5.0.0
matplotlib>=3.5.0

# Data utilities
tabulate>=0.9.0
python-dotenv>=0.20.0
certifi>=2022.0.0

# Machine Learning
scikit-learn>=1.0.0
torch>=2.0.0

# IoT / Sensors
paho-mqtt>=1.6.0

